refactor: extract JiraExportProvider+Networking (#638 slice A)
First foundational slice of #638. Moves the 7 auth/HTTP/JSON/error
helpers from JiraExportProvider.swift into a new
Sources/BugNarrator/Services/JiraExportProvider+Networking.swift
extension file, so subsequent slices B-E can reference them from
separate extension files.

Moved (byte-verbatim except for visibility widening):
- decodeJiraPayload<T>(_:from:description:) — private -> internal (widen)
- decodingFailureMessage(_:) — stays private (single-caller, moves with)
- basicAuthValue(email:apiToken:) — stays private (single-caller, moves with)
- jiraRequestURL(baseURL:path:queryItems:) — private -> internal (widen)
- authenticatedRequest(url:httpMethod:email:apiToken:includesJSONContentType:) — private -> internal (widen)
- mapJiraError(statusCode:data:configuration:retryAfterSeconds:) — private -> internal (widen)
- decodeJiraMessage(from:) — stays private (single-caller, moves with)
- JiraErrorResponse struct — file-scope private -> module-internal (widen)

Visibility widening rationale: the 4 widened helpers are called
from the base file (JiraExportProvider main class body). Cross-file
access requires at least internal in Swift; private on an extension
member is file-scoped. The 3 single-caller helpers (decodingFailure-
Message, basicAuthValue, decodeJiraMessage) stay private because
their sole callers also move into the same new extension file.

Codex spec-review flagged an AC clarification: 2/4 widened helpers
(jiraRequestURL, decodeJiraPayload) are single-cluster today (only
Discovery cluster calls them), not fan-out as the ticket claimed.
The widening is still required because Discovery callers stay in
the base file for this slice — implementation is unchanged.

External impact: zero. rg across Sources/ and Tests/ confirms
no external file references the widened symbols. Only same-name
identifier is `authenticatedRequest` on GitHubExportProvider (a
different actor).

Line-count delta:
- JiraExportProvider.swift: 1255 -> 1139 (-116 lines)
- JiraExportProvider+Networking.swift (new): 127 lines
- xcodeproj: +4 (xcodegen)

Test coverage: JiraExportProviderTests continues to pass unchanged.

Closes #874. Continues #638 (design sketch on parent).
